Hopefully fine-tuning SSH connection ignore for ssh-/gpg-agent

+++ zlogout
@@ -1,3 +1,3 @@
-if [ ! -z "$SSH_AGENT_PID" -a -z "$TMUX" ]; then
+if [ ! -z "$SSH_AGENT_PID" -a -z "$TMUX" -a -z "${SSH_CONNECTION:-}" ]; then
   kill $SSH_AGENT_PID
 fi
